+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
 - Updated onto-deside configuration 2025-04-09 (#202).
+- In the configuration file: per query `httpProxies` setting replaces global `httpProxy` setting (#4, #47).
 
 ## [1.7.0] - 2025-04-09